+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- Internals: The ability to store and change "naked" properties
+
 ## [0.6.2] - 2019-08-29
 
 ### Fixed